// vendor/expo-modules-autolinking/src/types.ts
export type PackageRevision = {
  name: string;
  path: string;
  version: string;
  config?: ExpoModuleConfig;
  duplicates?: PackageRevision[];
};

export type SearchResults = {
  [moduleName: string]: PackageRevision;
};
// ...
export interface CommonNativeModuleDescriptor {
  packageName: string;
  coreFeatures?: string[];
}
// ...
export interface ModuleWindowsConfig {
  name: string | null;
  class: string; // fully qualified: "ExpoCamera.CameraModule"
}

export interface ModuleDescriptorWindows extends CommonNativeModuleDescriptor {
  modules: ModuleWindowsConfig[];
  projectPath: string | null;
  debugOnly: boolean;
}